move schema validation to LibraryServiceSupport; move associated tests
