+++ src/test/scala/org/broadinstitute/dsde/firecloud/service/LibraryApiServiceSpec.scala
@@ -1,7 +1,5 @@
 package org.broadinstitute.dsde.firecloud.service
 
-import java.net.URL
-
 import org.broadinstitute.dsde.firecloud.Application
 import org.broadinstitute.dsde.firecloud.dataaccess._
 import org.broadinstitute.dsde.firecloud.mock.MockUtils
@@ -11,17 +9,9 @@ import org.broadinstitute.dsde.firecloud.webservice.LibraryApiService
 import org.mockserver.integration.ClientAndServer
 import org.mockserver.integration.ClientAndServer._
 import org.mockserver.model.HttpRequest._
-import org.parboiled.common.FileUtils
 
-import scala.util.Try
 import spray.http.HttpMethods
 import spray.http.StatusCodes._
-import spray.json._
-import spray.util._
-import org.everit.json.schema.Schema
-import org.everit.json.schema.loader.SchemaLoader
-import org.json.JSONObject
-import org.json.JSONTokener
 
 class LibraryApiServiceSpec extends BaseServiceSpec with LibraryApiService {
 
@@ -64,28 +54,5 @@ class LibraryApiServiceSpec extends BaseServiceSpec with LibraryApiService {
         }
       }
     }
-
-    "in its schema definition" - {
-      "has valid JSON" in {
-        val fileContents = FileUtils.readAllTextFromResource("library/attribute-definitions.json")
-        val jsonVal:Try[JsValue] = Try(fileContents.parseJson)
-        assert(jsonVal.isSuccess, "Schema should be valid json")
-      }
-      "has valid JSON Schema" in {
-        val schemaStream = new URL("http://json-schema.org/draft-04/schema").openStream()
-
-        try {
-          val fileContents = FileUtils.readAllTextFromResource("library/attribute-definitions.json")
-
-          val rawSchema:JSONObject = new JSONObject(new JSONTokener(schemaStream))
-          val schema:Schema = SchemaLoader.load(rawSchema)
-          schema.validate(new JSONObject(fileContents))
-
-        } finally {
-          schemaStream.close()
-        }
-      }
-    }
-
   }
 }
